refactor transformations
